+++ src/index.js
@@ -16,7 +16,7 @@ const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <BrowserRouter>
-      <MyPage/>
+      <ShoppingBasket/>
     </BrowserRouter>
   </React.StrictMode>
 );
